2_YouTube_NewVideo_Transcript/requirements.txt
python-dotenv
requests
youtube-transcript-api
